Move cmd/internal to subpackage
