+++ cmd/nerdctl/internal/internal_oci_hook.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package main
+package internal
 
 import (
 	"errors"
